Refactor files and props
