+++ frontend/scripts/react-components/shared/profile-selector/profile-panel-footer/profile-panel-footer.js
@@ -1,0 +1,3 @@
+export {
+  default
+} from 'react-components/shared/profile-selector/profile-panel-footer/profile-panel-footer.component';
