+++ frontend/scripts/react-components/shared/profile-selector/profile-panel-footer/profile-panel-footer.scss
