+++ frontend/scripts/react-components/shared/profile-selector/profile-panel/profile-panel.component.jsx
@@ -5,24 +5,18 @@ import 'react-components/shared/profile-selector/profile-panel/profile-panel.scs
 import BlockSwitch from 'react-components/shared/block-switch/block-switch.component';
 import Heading from 'react-components/shared/heading';
 import CommoditiesPanel from 'react-components/dashboard-element/dashboard-panel/commodities-panel.component';
-import ProfileStepPanel from 'react-components/shared/profile-selector/profile-panel/profile-step-panel.component';
+import ProfileStepPanel from 'react-components/shared/profile-selector/profile-panel/profile-step-panel';
 import getPanelStepName from 'utils/getProfilePanelName';
 
 function ProfilePanel(props) {
   const {
     step,
     setProfilesActiveItem,
-    clearProfilesPanel,
-    getSearchResults,
-    setProfilesSearchResult,
-    setProfilesActiveTab,
-    getMoreItems,
     profileType,
     blocks,
     setProfilesPage,
     panels,
     data,
-    tabs,
     loading
   } = props;
   const singularTypes = {
@@ -59,20 +53,7 @@ function ProfilePanel(props) {
             </Heading>
           </Heading>
           <div className="row profile-panel-content">
-            <ProfileStepPanel
-              profileType={profileType}
-              panelName={getPanelStepName(step)}
-              setProfilesActiveItem={setProfilesActiveItem}
-              setProfilesActiveTab={setProfilesActiveTab}
-              clearProfilesPanel={clearProfilesPanel}
-              getSearchResults={getSearchResults}
-              setSearchResult={setProfilesSearchResult}
-              getMoreItems={getMoreItems}
-              data={data}
-              loading={loading}
-              tabs={tabs}
-              panels={panels}
-            />
+            <ProfileStepPanel panelName={getPanelStepName(step)} />
           </div>
         </div>
       );
@@ -108,12 +89,6 @@ function ProfilePanel(props) {
 ProfilePanel.propTypes = {
   step: PropTypes.number,
   setProfilesActiveItem: PropTypes.func.isRequired,
-  clearProfilesPanel: PropTypes.func.isRequired,
-  setProfilesActiveTab: PropTypes.func.isRequired,
-  getSearchResults: PropTypes.func.isRequired,
-  setProfilesSearchResult: PropTypes.func.isRequired,
-  getMoreItems: PropTypes.func.isRequired,
-  tabs: PropTypes.array,
   profileType: PropTypes.string,
   blocks: PropTypes.array,
   setProfilesPage: PropTypes.func.isRequired,
