+++ frontend/scripts/react-components/shared/profile-selector/profile-panel/profile-step-panel/profile-step-panel.selectors.js
